server/requirements.txt
flask
flask-cors
transformers
Pillow
google-generativeai
python-dotenv
together
requests
PyPDF2
python-docx
spacy
opencv-python
matplotlib
insightface
scipy
torch
onnxruntime
numpy